image_types_tegra.bbclass: improve code signing server integration
Added settings:

TEGRA_SIGNING_EXCLUDE_TOOLS: when set to "1", the host-side
tools will not be copied into the working directory for signing,
to reduce the size of the package sent to the code signing server.
A signing server shouldn't be executing scripts or programs sent
to it anyway.

TEGRA_SIGNING_EXTRA_DEPS: for including additional dependencies
that might be required by the code signing client.

+++ classes/image_types_tegra.bbclass
@@ -9,6 +9,8 @@ INITRD_IMAGE ??= ""
 KERNEL_ARGS ??= ""
 TEGRA_SIGNING_ARGS ??= ""
 TEGRA_SIGNING_ENV ??= ""
+TEGRA_SIGNING_EXCLUDE_TOOLS ??= ""
+TEGRA_SIGNING_EXTRA_DEPS ??= ""
 
 DTBFILE ?= "${@os.path.basename(d.getVar('KERNEL_DEVICETREE').split()[0])}"
 LNXFILE ?= "${@'${IMAGE_UBOOT}-${MACHINE}.bin' if '${IMAGE_UBOOT}' != '' else '${INITRD_IMAGE}-${MACHINE}.cboot'}"
@@ -425,7 +427,9 @@ create_tegraflash_pkg_tegra186() {
 	    ln -sf $fname ./
 	done
     done
-    cp -R ${STAGING_BINDIR_NATIVE}/tegra186-flash/* .
+    if [ "${TEGRA_SIGNING_EXCLUDE_TOOLS}" != "1" ]; then
+        cp -R ${STAGING_BINDIR_NATIVE}/tegra186-flash/* .
+    fi
     dd if=/dev/zero of=badpage.bin bs=4096 count=1
     tegraflash_custom_pre
     mksparse -v --fillpattern=0 "${IMAGE_TEGRAFLASH_ROOTFS}" ${IMAGE_BASENAME}.img
@@ -438,6 +442,9 @@ END
     chmod +x doflash.sh
     tegraflash_custom_post
     tegraflash_custom_sign_pkg
+    if [ "${TEGRA_SIGNING_EXCLUDE_TOOLS}" = "1" ]; then
+        cp -R ${STAGING_BINDIR_NATIVE}/tegra186-flash/* .
+    fi
     rm -f ${IMGDEPLOYDIR}/${IMAGE_NAME}.tegraflash.zip
     zip -r ${IMGDEPLOYDIR}/${IMAGE_NAME}.tegraflash.zip .
     ln -sf ${IMAGE_NAME}.tegraflash.zip ${IMGDEPLOYDIR}/${IMAGE_LINK_NAME}.tegraflash.zip
@@ -502,7 +509,7 @@ do_image_tegraflash[depends] += "zip-native:do_populate_sysroot dtc-native:do_po
                                  virtual/kernel:do_deploy \
                                  ${@'${INITRD_IMAGE}:do_image_complete' if d.getVar('INITRD_IMAGE') != '' else  ''} \
                                  ${@'${IMAGE_UBOOT}:do_deploy ${IMAGE_UBOOT}:do_populate_lic' if d.getVar('IMAGE_UBOOT') != '' else  ''} \
-                                 ${TEGRAFLASH_DEPS_EXTRA}"
+                                 ${TEGRAFLASH_DEPS_EXTRA} ${TEGRA_SIGNING_EXTRA_DEPS}"
 IMAGE_TYPEDEP_tegraflash += "${IMAGE_TEGRAFLASH_FS_TYPE}"
 
 oe_make_bup_payload() {
@@ -582,32 +589,34 @@ oe_make_bup_payload_common() {
     rm -f ./slot_metadata.bin
     cp ${STAGING_DATADIR}/tegraflash/slot_metadata.bin ./
     mkdir ./rollback
-    ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/rollback_parser.py ./rollback/
     ln -snf ${STAGING_DATADIR}/nv_tegra/rollback/t${@d.getVar('NVIDIA_CHIP')[2:]}x ./rollback/
-    ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/${SOC_FAMILY}-flash-helper.sh ./
-    sed -e 's,^function ,,' ${STAGING_BINDIR_NATIVE}/tegra186-flash/l4t_bup_gen.func > ./l4t_bup_gen.func
-    ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/*.py .
-    rm -f ./doflash.sh
-    cat <<EOF > ./doflash.sh
+    if [ "${TEGRA_SIGNING_EXCLUDE_TOOLS}" != "1" ]; then
+        ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/rollback_parser.py ./rollback/
+        ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/${SOC_FAMILY}-flash-helper.sh ./
+        sed -e 's,^function ,,' ${STAGING_BINDIR_NATIVE}/tegra186-flash/l4t_bup_gen.func > ./l4t_bup_gen.func
+        ln -sf ${STAGING_BINDIR_NATIVE}/tegra186-flash/*.py .
+        rm -f ./doflash.sh
+        cat <<EOF > ./doflash.sh
 export BOARDID=${TEGRA_BOARDID}
 export FAB=${TEGRA_FAB}
 export fuselevel=fuselevel_production
 export localbootfile=${LNXFILE}
 EOF
-    if [ "${SOC_FAMILY}" = "tegra194" ]; then
-        cat <<EOF >> ./doflash.sh
+        if [ "${SOC_FAMILY}" = "tegra194" ]; then
+            cat <<EOF >> ./doflash.sh
 export CHIPREV=${TEGRA_CHIPREV}
 export BOARDSKU=${TEGRA_BOARDSKU}
 export BOARDREV=${TEGRA_BOARDREV}
 EOF
-        sdramcfg="${MACHINE}.cfg,${MACHINE}-override.cfg"
-    else
-        sdramcfg="${MACHINE}.cfg"
-    fi
-    cat <<EOF >>./doflash.sh
+            sdramcfg="${MACHINE}.cfg,${MACHINE}-override.cfg"
+        else
+            sdramcfg="${MACHINE}.cfg"
+        fi
+        cat <<EOF >>./doflash.sh
 MACHINE=${MACHINE} ./${SOC_FAMILY}-flash-helper.sh --bup ./flash.xml.in ${DTBFILE} $sdramcfg ${ODMDATA} "\$@"
 EOF
-    chmod +x ./doflash.sh
+        chmod +x ./doflash.sh
+    fi
     tegraflash_custom_sign_bup
     cd "$oldwd"
 }
@@ -624,6 +633,6 @@ CONVERSIONTYPES += "bup-payload"
 BUP_DEPS_EXTRA = ""
 BUP_DEPS_EXTRA_tegra186 = "virtual/secure-os:do_deploy"
 BUP_DEPS_EXTRA_tegra194 = "virtual/secure-os:do_deploy"
-CONVERSION_DEPENDS_bup-payload = "tegra186-flashtools-native tegra-bootfiles tegra186-redundant-boot nv-tegra-release dtc-native virtual/bootloader:do_deploy virtual/kernel:do_deploy ${BUP_DEPS_EXTRA}"
+CONVERSION_DEPENDS_bup-payload = "tegra186-flashtools-native tegra-bootfiles tegra186-redundant-boot nv-tegra-release dtc-native virtual/bootloader:do_deploy virtual/kernel:do_deploy ${BUP_DEPS_EXTRA} ${TEGRA_SIGNING_EXTRA_DEPS}"
 CONVERSION_CMD_bup-payload = "create_bup_payload_image ${type}"
 IMAGE_TYPES += "cpio.gz.cboot.bup-payload"
